# Made many `Rect/FRect` methods fastcall